Refactored user login/registration into Users app
