+++ ecommerce/auctions/static/users/styles.css
@@ -1,0 +1,270 @@
+* {
+  box-sizing: border-box;
+}
+
+body {
+  font-family: sans-serif;
+  font-size: 1.4rem;
+  display: flex;
+  justify-content: center;
+  width: 100%;
+  height: 100%;
+  margin: auto;
+}
+
+h2 {
+  margin: .5rem 0;
+}
+
+ul {
+  list-style: none;
+}
+
+li {
+  margin-left: 1.5rem;
+}
+
+p {
+  text-wrap: wrap;
+  word-wrap: break-word;
+}
+
+a {
+  text-decoration: none;
+  font-size: 1.7rem;
+  color: blue;
+  font-weight: bold;
+}
+
+a:hover {
+  transition: all .3s ease-in;
+  color: red;
+  text-decoration: underline;
+}
+
+.container {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  max-width: 1200px;
+  height: 100%;
+}
+
+.nav {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  width: 100%;
+  justify-content: space-between;
+}
+
+.nav-left {
+  display: flex;
+  flex-direction: column;
+}
+
+.nav-right {
+  display: flex;
+  flex-direction: row;
+}
+
+.index-container {
+  display: flex;
+  flex-direction: row;
+  max-width: 1200px;
+  flex-wrap: wrap;
+  height: 100%;
+  gap: 3rem;
+  
+}
+
+.item-container {
+  
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-end;
+  margin-top: 1rem;
+  width: 30%;
+  
+}
+
+.item-container img {
+  max-width: 80%;
+  max-height: 20rem;
+}
+
+.add-listing form {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  width: 50%;
+}
+
+.info-container {
+  font-size: 1.2rem;
+}
+
+.info-container a {
+  font-size: 1.2rem;
+}
+
+.info-container p{
+  margin: .2rem 0;
+}
+
+hr {
+  border-top: 1px solid gray;
+  margin: 1px 0;
+}
+
+.item-end-date {
+  margin-left: .2rem;
+  font-size: 1rem;
+  display: flex;
+  align-items: center;
+  gap: .4rem;
+}
+
+.info-container .item-name {
+  font-size: 1.5rem;
+  font-weight: bolder;
+  margin-top: .2rem;
+}
+
+.img-container {
+  display: flex;
+  justify-content: center;
+  margin-bottom: 2rem;
+}
+
+.img-container a{
+  display: flex;
+  justify-content: center;
+}
+
+.img-container img:hover {
+  transform: scale(1.2);
+}
+
+.header {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.form-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  align-self: center;
+  margin-top: 15rem;
+  border: 1px solid black;
+  box-shadow: 3px 3px 15px 1px rgba(0,0,0,0.8);
+  border-radius: .75rem;
+  padding: 1rem;
+}
+
+label{
+  font-size: 1.2rem;
+}
+
+
+input{
+  font-size: 1.4rem;
+  border-radius: .25rem;
+  margin-bottom: .5rem;
+  padding: .2rem;
+}
+
+.form-layout{
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+#login-header {
+  align-self: flex-start;
+  margin: 0 0 .5rem;
+}
+
+#register-link {
+  margin-top: .2rem;
+  font-size: 1.4rem;
+}
+
+#register-header {
+  align-self: flex-start;
+  margin: 0 0 .5rem;
+}
+
+.form-layout button {
+  width:80%;
+  margin: .5rem;
+  height: 2rem;
+  font-size: 1.5rem;
+  background-color: blue;
+  color: white;
+  border-radius: .5rem;
+}
+
+.form-layout button:hover {
+  box-shadow: 3px 3px 8px 3px rgba(0,0,0,0.5);
+}
+
+#category-select {
+  display: flex;
+  align-items: center;
+  gap: .5rem;
+}
+
+#category-select select {
+  font-size: 1.4rem;
+}
+
+#category-select button {
+  height: 2rem;
+  font-size: 1.2rem;
+}
+
+.listing-container {
+  align-self:center;
+  
+  width: 100%;
+  max-width: 30rem;
+}
+
+.listing-container .item-container {
+  margin-top: 3rem;
+  width: 100%;
+}
+
+.listing-container button{
+  height: 2rem;
+  font-size: 1.2rem;
+  border-radius: .25rem;
+}
+
+#bid-form {
+  margin: 1rem 0;
+  display: flex;
+  flex-direction: column;
+  max-width: 10rem;
+}
+
+#comment-form {
+  margin: 1rem 0;
+  display: flex;
+  flex-direction: column;
+  
+}
+
+#comment-form button {
+  margin-top: 1rem;
+  max-width: 15rem;
+}
+
+#comment-form textarea {
+  height: 10rem;
+  resize: none;
+}
