+++ ecommerce/auctions/templates/auctions/add_listing.html
@@ -1,7 +1,7 @@
 {% extends "auctions/layout.html" %}
 
 {% block body %}
-    <h2>Active Listings</h2>
+    <h2>Add Listing</h2>
     {% if message %}
       <p>{{ message }}</p>
     {% endif %}
