+++ ecommerce/auctions/urls.py
@@ -3,9 +3,6 @@ from . import views
 
 urlpatterns = [
     path("", views.index, name="index"),
-    path("login", views.login_view, name="login"),
-    path("logout", views.logout_view, name="logout"),
-    path("register", views.register, name="register"),
     path("listing/<int:id>", views.listing, name="listing"),
     path("category/<str:category_name>", views.category, name="category"),
     path("add_listing", views.add_listing, name="add_listing"),
